refactor horizontal partition tree
