+++ front-end/src/utils/horizontalTree.ts
@@ -1,45 +1,19 @@
-// import * as d3 from 'd3';
-import React from 'react';
-import { v4 as uuidv4 } from 'uuid';
+
 import * as d3 from 'd3';
 import { valueAccessor } from 'src/accessors/squareAccessors';
 import { dimension } from 'src/interfaces/interfaces';
-import { getColorDataAccessor, getCGenerator } from 'src/utils/treeAccess';
+import { sizeAccesorMin } from 'src/accessors/treeAccessors'
 
-export const parseOmitedLinks = (data: any) => {
-    return data.map((d: any) => <text
-        key={uuidv4()}
-        x={d.target.y + (d.source.y - d.target.y) / 2}
-        y={d.target.x + (d.source.x - d.target.x) / 2}
-        textAnchor="middle"
-        className="omitted-label"
-    >
-        {d.version}
-    </text>)
-}
 
-export const parseOmitedParitionLinks = (data: any) => {
-    return data.map((d: any) => <text
-        key={uuidv4()}
-        x={d.target.y + (d.source.y - d.target.y) / 2}
-        y={d.target.x + (d.source.x - d.target.x) / 2}
-        textAnchor="middle"
-        className="omitted-label"
-    >
-        {d.version}
-    </text>)
-}
+//get all the ommited
+const filterOmmited = (d: any) => d.data.omitted === true;
+const findNode = (node: any) => ((d: any) =>
+    d.data.groupId === node.data.groupId
+    && d.data.artifactId === node.data.artifactId
+    && d.data.omitted === false)
 
-//returns all the links of the ommited
-export const getOmmitedLinks = (nodes: any) => {
-    //get all the ommited
-    const filterOmmited = (d: any) => d.data.omitted === true;
-    const findNode = (node: any) => ((d: any) =>
-        d.data.groupId === node.data.groupId
-        && d.data.artifactId === node.data.artifactId
-        && d.data.omitted === false)
-    //get the target for each of the ommited
-    const createLinks = (linksMap: any, node: any) => {
+const getLinks = (nodes: any) => {
+    return (linksMap: any, node: any) => {
         const parent = node.parent;
         const replacement = nodes.find(findNode(node));
         const link = {
@@ -50,9 +24,17 @@ export const getOmmitedLinks = (nodes: any) => {
         //if either of the source or target are not visible then it should not pain it
         return (parent.data.deleted === false && replacement.data.deleted === false) || (parent.data.visible === false && replacement.data.visible === false) ? [...linksMap, link] : [...linksMap];
     }
+}
+
+//returns all the links of the ommited
+export const getOmmitedLinks = (filtered: any) => {
+    const nodes = filtered
+        .sum(sizeAccesorMin)
+        .descendants()
+    //get the target for each of the ommited
     return nodes
         .filter(filterOmmited)
-        .reduce(createLinks, [])
+        .reduce(getLinks(nodes), [])
 }
 
 export const getParitionTree = (size: number[], padding: number) => {
@@ -169,8 +151,3 @@ export const getNodesFromParitionTree = (dimensions: dimension, sizeAccesorMin: 
         .map(addNewSize(heightPercent, 80, dimensions.boundedHeight))
 }
 
-export const getColor = (colorSelected: "NONE" | "DEPENDENCY_TYPE" | "USAGE_RATIO" | "GROUP_ID" | "TRANSPARENT", nodes: any[]) => {
-    const colorDataAccessor: (d: any) => string = getColorDataAccessor(colorSelected)
-    const colorGenerator: any = getCGenerator(colorSelected, nodes);
-    return (d: any) => colorGenerator(colorDataAccessor(d));
-}
